# swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "description": "API for Social Media",
      "title": "Social-media App"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "components": {
      "securitySchemes": {
        "JWT": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
        }
      }
    },
    "paths": {
...
      "/api/posts": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get User's Posts",
          "description": "Retrieve posts for the logged-in user",
          "security": [
            {
              "JWT": []
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "example": [
                    {
                      "id": 1,
                      "userId": 4,
                      "caption": "Example Post",
                      "imageUrl": "http://example.com/image.jpg",
                      "comments": [
                        {
                          "userId": 5,
                          "comment": "Nice post!"
                        }
                      ]
                    }
                  ]
                }
              }
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "post": {
          "tags": ["Posts"],
          "summary": "Add a Post",
          "description": "Create a new post",
          "security": [
            {
              "JWT": []
            }
          ],
          "requestBody": {
            "description": "Post details",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caption": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "required": ["caption", "image"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Post created successfully",
              "content": {
                "application/json": {
                  "example": {
                    "userId": "123",
                    "caption": "Example Post",
                    "image": "example.jpg"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["Posts"],
          "summary": "Update a Post",
          "description": "Update an existing post",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the post to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Updated post details",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caption": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }